# INFRA-504: Replace out of process nodes with in process nodes