--- a/node/src/integration-test-slow/kotlin/net/corda/node/utilities/registration/NodeRegistrationTest.kt
+++ b/node/src/integration-test-slow/kotlin/net/corda/node/utilities/registration/NodeRegistrationTest.kt
@@ -17,6 +17,7 @@ import net.corda.testing.common.internal.testNetworkParameters
 import net.corda.testing.core.SerializationEnvironmentRule
 import net.corda.testing.driver.internal.incrementalPortAllocation
 import net.corda.coretesting.internal.DEV_ROOT_CA
+import net.corda.testing.driver.NodeParameters
 import net.corda.testing.node.NotarySpec
 import net.corda.testing.node.internal.SharedCompatibilityZoneParams
 import net.corda.testing.node.internal.internalDriver
@@ -85,9 +86,13 @@ class NodeRegistrationTest {
                 compatibilityZone = compatibilityZone,
                 notarySpecs = listOf(NotarySpec(notaryName)),
                 notaryCustomOverrides = mapOf("devMode" to false),
+                startNodesInProcess = true,
                 allowHibernateToManageAppSchema = false
         ) {
-            startNode(providedName = aliceName, customOverrides = mapOf("devMode" to false)).getOrThrow()
+            startNode(NodeParameters(
+                    providedName = aliceName,
+                    customOverrides = mapOf("devMode" to false)
+            )).getOrThrow()
 
             assertThat(registrationHandler.idsPolled).containsOnly(
                     aliceName.organisation,