--- a/node/src/integration-test/kotlin/net/corda/node/CordappScanningDriverTest.kt
+++ b/node/src/integration-test/kotlin/net/corda/node/CordappScanningDriverTest.kt
@@ -13,6 +13,7 @@ import net.corda.testing.core.ALICE_NAME
 import net.corda.testing.core.BOB_NAME
 import net.corda.testing.core.singleIdentity
 import net.corda.testing.driver.DriverParameters
+import net.corda.testing.driver.NodeParameters
 import net.corda.testing.driver.driver
 import net.corda.testing.node.User
 import net.corda.testing.node.internal.enclosedCordapp
@@ -24,10 +25,19 @@ class CordappScanningDriverTest {
 	fun `sub-classed initiated flow pointing to the same initiating flow as its super-class`() {
         val user = User("u", "p", setOf(startFlow<ReceiveFlow>()))
         // The driver will automatically pick up the annotated flows below
-        driver(DriverParameters(notarySpecs = emptyList(), cordappsForAllNodes = listOf(enclosedCordapp()))) {
+        driver(DriverParameters(
+                notarySpecs = emptyList(),
+                startNodesInProcess = true,
+                cordappsForAllNodes = listOf(enclosedCordapp())
+        )) {
             val (alice, bob) = listOf(
-                    startNode(providedName = ALICE_NAME, rpcUsers = listOf(user)),
-                    startNode(providedName = BOB_NAME)).transpose().getOrThrow()
+                    startNode(NodeParameters(
+                            providedName = ALICE_NAME,
+                            rpcUsers = listOf(user)
+                    )),
+                    startNode(NodeParameters(
+                            providedName = BOB_NAME
+                    ))).transpose().getOrThrow()
             val initiatedFlowClass = CordaRPCClient(alice.rpcAddress)
                     .start(user.username, user.password)
                     .proxy