Make tests use in process nodes wherever possible in the hopes of a substantial speed increase.

NodeRegistrationTest.`node registration correct root cert`
 - Before: 45s 252ms (first run) | 45s 666ms (second run)
 - After: 18s 820ms (first run) | 17S 850ms (second run)
**~60% speedup**

GenericReturnFailureReproductionIntegrationTest.flowShouldReturnGenericList
 - Before: 18s 996ms (first run) | 18s 560ms (second run)
 - After: 22s 585ms (first run) | 18s 818ms (second run)
**No real difference on this one outside of jitter**

CordappScanningDriverTest.`sub-classed initiated flow pointing to the same initiating flow as its super-class`
 - Before: 26s 964ms (first run) | 28s 500ms (second run)
 - After: 22s 689ms (first run) | 23s 162ms (second run)
**~18% Speedup**