@rnicoll 2020-08-05
We've actually been just talking in the leads meeting about how issues with https://r3-cev.atlassian.net/browse/CORDA-3933 get worse with in-process nodes, so may need to hold off on squashing this until we can get the serialization environment under control, at risk of making things less stable.