@vkolomeyko 2020-10-08
CI reports failure of `test that when states are being spent at the same time that schedules trigger everything is processed – net.corda.node.services.events.ScheduledFlowIntegrationTests` which probably needs investigation.