@rnicoll 2020-10-08
Running the tests a few times to confirm stability @roastario but seems okay so far